+++ Pureisuteshon.UI/Models/MessageGroups/MessageGroup.cs
@@ -31,7 +31,7 @@ namespace PlayStation_App.Models.MessageGroups
         public ThumbnailDetail ThumbnailDetail { get; set; }
 
         [JsonProperty("messageUid")]
-        public int MessageUid { get; set; }
+        public float MessageUid { get; set; }
 
         [JsonProperty("sentMessageId")]
         public string SentMessageId { get; set; }
