+++ Pureisuteshon.UI/Models/Messages/SentMessageResponse.cs
@@ -16,7 +16,7 @@ namespace PlayStation_App.Models.Messages
         public string MessageGroupModifiedDate { get; set; }
 
         [JsonProperty("messageUid")]
-        public int MessageUid { get; set; }
+        public float MessageUid { get; set; }
 
         [JsonProperty("sentMessageId")]
         public string SentMessageId { get; set; }
